pyproject.toml
[project]
name = "forge-plugin"
version = "6.0.0"
description = "Claude Code plugin: 10-stage autonomous pipeline (forge)"
requires-python = ">=3.10"
readme = "README.md"
license = { text = "Proprietary" }

[tool.setuptools]
# Metadata-only project: forge ships its tools as repo-rooted scripts/modules
# (under hooks/, shared/, tests/), not as an importable package. Setuptools'
# auto-discovery would otherwise pick up multiple top-level dirs and refuse.
py-modules = []

[project.optional-dependencies]
otel = [
  "opentelemetry-api>=1.41.0",
  "opentelemetry-sdk>=1.41.0",
  "opentelemetry-exporter-otlp>=1.41.0",
]
test = [
  "pydantic>=2.0",
  "pyyaml>=6.0",
  "pytest>=8.0",
  "jsonschema>=4.0.0",
  "tree-sitter>=0.25.2,<0.26",
  "tree-sitter-language-pack>=1.6.2,<2.0",
  "numpy>=1.26",
  "ruff>=0.5",
  "mypy>=1.0",
  "types-jsonschema>=4.0.0",
  "types-PyYAML>=6.0",
  "tomli>=2.0; python_version < '3.11'",
]

[tool.ruff]
target-version = "py310"
line-length = 100
extend-exclude = ["modules/", "agents/", "skills/", "docs/"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM"]
ignore = ["E501"]  # line length handled elsewhere

[tool.mypy]
python_version = "3.10"

[[tool.mypy.overrides]]
module = "tests.evals.benchmark.*"
disallow_untyped_defs = true
strict_optional = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests/unit", "tests/contract", "tests/scenario", "tests/integration"]
python_files = ["test_*.py", "*_test.py"]
pythonpath = ["."]
addopts = "-ra"
